server/lib/nl/detection/agent/requirements.txt
google-adk
pydantic